+++ packages/hooks/use-counter/package.json
