+++ packages/hooks/use-interval/package.json
