+++ packages/hooks/use-switch/src/switch.ts
