+++ packages/hooks/use-switch/src/switch.utils.ts
